requirements.txt
streamlit==1.38.0
rich==13.7.1  # 儘量壓在 <14，避免 Cloud 後步驟升到 14 造成不相容
# 改用 ONNXRuntime（CPU）避免 Py3.13 下 torch 相關安裝失敗
onnxruntime>=1.18.0
numpy>=1.26.4,<3.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.1
tqdm==4.66.4
requests==2.32.3
